EXPERIMENTAL: Update types for `withRouter`
In an attempt to make `withRouter` more friendly to decorators, I've
added some additional typings and tests to `next/router`.  I'm leaving
this as *experimental* for now, because I'm not completely thrilled with
the things that I've left as `TODO`s.  If you have ideas on how to
correct them (without loosening the type constraints), feel free to
submit a pull request against this change.